Changes and bug fixes
- Join functionality added
- Remove the custom error page while delete the table and project and
  redirect to homepage and displayed the message there
- Fixed the issues in Merge functionality
- Created a common function to create and insert data for dynamic
  tables(datatype & dashboard)
- Created a common function to get the column names of a table
- Images added for join functionality

Committed datetime: 07 June 2024 at 10:15 pm 😔

+++ src/sidebar.php
@@ -7,20 +7,26 @@ include_once "constants/common_constants.php";
 
 $userNotificationMsg = '';
 $notificationClassName = '';
-if ($_SERVER["REQUEST_METHOD"] == "POST") {
+
+if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['formName']) && $_POST['formName'] == 'sidebarForm') {
     include_once("utilities/common_utils.php");
     $projectId = $_POST['deleteProjectId'] ?? 0;
     $tableId = $_POST['deleteTableId'] ?? 0;
 
     if($projectId > 0) {
         $response = deleteAllProjectRelatedData($projectId, $pdo);
+        $actionType = 'project';
     } else if($tableId > 0) {
         $response = deleteAllTableRelatedData($tableId, $pdo);
+        $actionType = 'table';
     }
 
     $notificationClassName = $response['notificationClassName'] ?? '';
     $userNotificationMsg = $response['userNotificationMsg'] ?? '';
+
+    header("Location:home.php?msg=$actionType");
 }
+
 $stmt = $pdo->prepare("SELECT id, name FROM projects");
 $stmt->execute();
 $projects = $stmt->fetchAll(PDO::FETCH_ASSOC);
